fix: wrong digest attack
This commit fix a wrong digest vulnerability in `sqisign.Verify` function.

The SQISIGN `crypto_sign_open` function is meant to recover the original message (m) from the signed message (sm) and its signature. It does not check if the recovered message matches the provided digest. It only validates the signature against the signed message.

In other words, before this fix, any valid signature could have been be used to validate a non-valid digest.