fix: PublicKey.Verify buffer overflow
The previous implementation of `PublicKey.Verify` was vulnerable to a buffer overflow. It allocate the buffer based on the size of the caller-provided `digest` argument.

If the actual message inside the signature was larger than the provided `digest`, the underlying C function would write past the allocated buffer, resulting in a SIGABRT.

The new implementation now uses the known property of the API instead:
- `message_length = signed_message_length - signature_overhead`.